requirements.txt
discord.py>=2.4.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
redis>=5.2.0,<6.0.0
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
python-dotenv>=1.1.0,<2.0.0
